+++ crates/cli/cli-components/src/traits/bootstrap.rs
@@ -1,3 +1,5 @@
+use cgp::core::component::WithProvider;
+use cgp::core::types::traits::ProvideType;
 use cgp::prelude::*;
 
 #[derive_component(BootstrapTypeComponent, ProvideBootstrapType<App>)]
@@ -9,4 +11,13 @@ pub trait HasBootstrapType: Async {
 #[async_trait]
 pub trait CanLoadBootstrap<Args: Async>: HasBootstrapType + HasErrorType {
     async fn load_bootstrap(&self, args: &Args) -> Result<Self::Bootstrap, Self::Error>;
+}
+
+impl<App, Provider, Bootstrap> ProvideBootstrapType<App> for WithProvider<Provider>
+where
+    App: Async,
+    Bootstrap: Async,
+    Provider: ProvideType<BootstrapTypeComponent, App, Type = Bootstrap>,
+{
+    type Bootstrap = Bootstrap;
 }