+++ crates/cosmos/cosmos-integration-tests/src/traits/bootstrap/mod.rs
@@ -1,5 +1,4 @@
 pub mod build_chain;
 pub mod compat_mode;
 pub mod cosmos_builder;
-pub mod gas_denom;
 pub mod relayer_chain_config;