+++ crates/cosmos/cosmos-test-components/src/bootstrap/impls/modifiers/mod.rs
@@ -1,2 +1,3 @@
 pub mod no_modify_comet_config;
+pub mod no_modify_cosmos_sdk_config;
 pub mod no_modify_genesis_config;