+++ crates/cosmos/cosmos-test-components/src/bootstrap/traits/modifiers/mod.rs
@@ -1,2 +1,3 @@
 pub mod modify_comet_config;
+pub mod modify_cosmos_sdk_config;
 pub mod modify_genesis_config;