+++ crates/cosmos/cosmos-wasm-relayer/src/context/cosmos_bootstrap.rs
@@ -1,4 +1,7 @@
 use alloc::sync::Arc;
+use cgp::core::component::UseContext;
+use hermes_cosmos_test_components::bootstrap::impls::modifiers::no_modify_cosmos_sdk_config::NoModifyCosmosSdkConfig;
+use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_cosmos_sdk_config::CosmosSdkConfigModifierComponent;
 use std::path::PathBuf;
 
 use cgp::core::error::{ErrorRaiserComponent, ErrorTypeComponent};
@@ -8,9 +11,10 @@ use hermes_cosmos_integration_tests::impls::bootstrap::build_cosmos_chain_driver
 use hermes_cosmos_integration_tests::impls::bootstrap::relayer_chain_config::BuildRelayerChainConfig;
 use hermes_cosmos_integration_tests::impls::bootstrap::types::ProvideCosmosBootstrapChainTypes;
 use hermes_cosmos_integration_tests::traits::bootstrap::build_chain::ChainBuilderWithNodeConfigComponent;
-use hermes_cosmos_integration_tests::traits::bootstrap::compat_mode::CompatModeGetter;
-use hermes_cosmos_integration_tests::traits::bootstrap::cosmos_builder::CosmosBuilderGetter;
-use hermes_cosmos_integration_tests::traits::bootstrap::gas_denom::GasDenomGetter;
+use hermes_cosmos_integration_tests::traits::bootstrap::compat_mode::{
+    CompatModeGetterComponent, UseCompatMode37,
+};
+use hermes_cosmos_integration_tests::traits::bootstrap::cosmos_builder::CosmosBuilderGetterComponent;
 use hermes_cosmos_integration_tests::traits::bootstrap::relayer_chain_config::RelayerChainConfigBuilderComponent;
 use hermes_cosmos_relayer::contexts::build::CosmosBuilder;
 use hermes_cosmos_test_components::bootstrap::components::cosmos_sdk::*;
@@ -18,13 +22,11 @@ use hermes_cosmos_test_components::bootstrap::impls::generator::wallet_config::G
 use hermes_cosmos_test_components::bootstrap::impls::modifiers::no_modify_comet_config::NoModifyCometConfig;
 use hermes_cosmos_test_components::bootstrap::impls::modifiers::no_modify_genesis_config::NoModifyGenesisConfig;
 use hermes_cosmos_test_components::bootstrap::traits::chain::build_chain_driver::ChainDriverBuilderComponent;
-use hermes_cosmos_test_components::bootstrap::traits::fields::account_prefix::AccountPrefixGetter;
-use hermes_cosmos_test_components::bootstrap::traits::fields::chain_command_path::ChainCommandPathGetter;
-use hermes_cosmos_test_components::bootstrap::traits::fields::chain_store_dir::ChainStoreDirGetter;
-use hermes_cosmos_test_components::bootstrap::traits::fields::denom::{
-    DenomForStaking, DenomForTransfer, DenomPrefixGetter,
-};
-use hermes_cosmos_test_components::bootstrap::traits::fields::random_id::RandomIdFlagGetter;
+use hermes_cosmos_test_components::bootstrap::traits::fields::account_prefix::AccountPrefixGetterComponent;
+use hermes_cosmos_test_components::bootstrap::traits::fields::chain_command_path::ChainCommandPathGetterComponent;
+use hermes_cosmos_test_components::bootstrap::traits::fields::chain_store_dir::ChainStoreDirGetterComponent;
+use hermes_cosmos_test_components::bootstrap::traits::fields::denom::DenomPrefixGetterComponent;
+use hermes_cosmos_test_components::bootstrap::traits::fields::random_id::RandomIdFlagGetterComponent;
 use hermes_cosmos_test_components::bootstrap::traits::generator::generate_wallet_config::WalletConfigGeneratorComponent;
 use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_comet_config::CometConfigModifierComponent;
 use hermes_cosmos_test_components::bootstrap::traits::modifiers::modify_genesis_config::CosmosGenesisConfigModifierComponent;
@@ -39,9 +41,8 @@ use hermes_test_components::driver::traits::types::chain_driver::ChainDriverType
 use hermes_wasm_test_components::impls::bootstrap::build_chain_driver::BuildChainDriverAndInitWasmClient;
 use hermes_wasm_test_components::impls::bootstrap::genesis_config::ModifyWasmGenesisConfig;
 use hermes_wasm_test_components::impls::bootstrap::node_config::ModifyWasmNodeConfig;
-use hermes_wasm_test_components::traits::bootstrap::client_byte_code::WasmClientByteCodeGetter;
-use hermes_wasm_test_components::traits::bootstrap::gov_authority::GovernanceProposalAuthorityGetter;
-use ibc_relayer::config::compat_mode::CompatMode;
+use hermes_wasm_test_components::traits::bootstrap::client_byte_code::WasmClientByteCodeGetterComponent;
+use hermes_wasm_test_components::traits::bootstrap::gov_authority::GovernanceProposalAuthorityGetterComponent;
 
 /**
    A bootstrap context for bootstrapping a new Cosmos chain, and builds
@@ -50,13 +51,13 @@ use ibc_relayer::config::compat_mode::CompatMode;
 #[derive(HasField)]
 pub struct CosmosWithWasmClientBootstrap {
     pub runtime: HermesRuntime,
-    pub builder: Arc<CosmosBuilder>,
+    pub cosmos_builder: Arc<CosmosBuilder>,
     pub should_randomize_identifiers: bool,
     pub chain_store_dir: PathBuf,
     pub chain_command_path: PathBuf,
     pub account_prefix: String,
-    pub staking_denom: String,
-    pub transfer_denom: String,
+    pub staking_denom_prefix: String,
+    pub transfer_denom_prefix: String,
     pub wasm_client_byte_code: Vec<u8>,
     pub governance_proposal_authority: String,
 }
@@ -92,6 +93,21 @@ delegate_components! {
             ChainDriverTypeComponent,
         ]:
             ProvideCosmosBootstrapChainTypes,
+        [
+            ChainStoreDirGetterComponent,
+            ChainCommandPathGetterComponent,
+            AccountPrefixGetterComponent,
+            DenomPrefixGetterComponent,
+            RandomIdFlagGetterComponent,
+            WasmClientByteCodeGetterComponent,
+            GovernanceProposalAuthorityGetterComponent,
+            CosmosBuilderGetterComponent,
+        ]:
+            UseContext,
+        CompatModeGetterComponent:
+            UseCompatMode37,
+        CosmosSdkConfigModifierComponent:
+            NoModifyCosmosSdkConfig,
         RelayerChainConfigBuilderComponent:
             BuildRelayerChainConfig,
         ChainBuilderWithNodeConfigComponent:
@@ -102,89 +118,5 @@ delegate_components! {
             ModifyWasmGenesisConfig<NoModifyGenesisConfig>,
         CometConfigModifierComponent:
             ModifyWasmNodeConfig<NoModifyCometConfig>,
-    }
-}
-
-impl ChainStoreDirGetter<CosmosWithWasmClientBootstrap>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn chain_store_dir(bootstrap: &CosmosWithWasmClientBootstrap) -> &PathBuf {
-        &bootstrap.chain_store_dir
-    }
-}
-
-impl ChainCommandPathGetter<CosmosWithWasmClientBootstrap>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn chain_command_path(bootstrap: &CosmosWithWasmClientBootstrap) -> &PathBuf {
-        &bootstrap.chain_command_path
-    }
-}
-
-impl RandomIdFlagGetter<CosmosWithWasmClientBootstrap> for CosmosWithWasmClientBootstrapComponents {
-    fn should_randomize_identifiers(bootstrap: &CosmosWithWasmClientBootstrap) -> bool {
-        bootstrap.should_randomize_identifiers
-    }
-}
-
-impl DenomPrefixGetter<CosmosWithWasmClientBootstrap, DenomForStaking>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn denom_prefix(bootstrap: &CosmosWithWasmClientBootstrap, _label: DenomForStaking) -> &str {
-        &bootstrap.staking_denom
-    }
-}
-
-impl DenomPrefixGetter<CosmosWithWasmClientBootstrap, DenomForTransfer>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn denom_prefix(bootstrap: &CosmosWithWasmClientBootstrap, _label: DenomForTransfer) -> &str {
-        &bootstrap.transfer_denom
-    }
-}
-
-impl AccountPrefixGetter<CosmosWithWasmClientBootstrap>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn account_prefix(bootstrap: &CosmosWithWasmClientBootstrap) -> &str {
-        &bootstrap.account_prefix
-    }
-}
-
-impl CompatModeGetter<CosmosWithWasmClientBootstrap> for CosmosWithWasmClientBootstrapComponents {
-    fn compat_mode(_bootstrap: &CosmosWithWasmClientBootstrap) -> Option<&CompatMode> {
-        const COMPAT_MODE: CompatMode = CompatMode::V0_37;
-
-        Some(&COMPAT_MODE)
-    }
-}
-
-impl GasDenomGetter<CosmosWithWasmClientBootstrap> for CosmosWithWasmClientBootstrapComponents {
-    fn gas_denom(bootstrap: &CosmosWithWasmClientBootstrap) -> &str {
-        &bootstrap.staking_denom
-    }
-}
-
-impl CosmosBuilderGetter<CosmosWithWasmClientBootstrap>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn cosmos_builder(bootstrap: &CosmosWithWasmClientBootstrap) -> &CosmosBuilder {
-        &bootstrap.builder
-    }
-}
-
-impl WasmClientByteCodeGetter<CosmosWithWasmClientBootstrap>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn wasm_client_byte_code(bootstrap: &CosmosWithWasmClientBootstrap) -> &Vec<u8> {
-        &bootstrap.wasm_client_byte_code
-    }
-}
-
-impl GovernanceProposalAuthorityGetter<CosmosWithWasmClientBootstrap>
-    for CosmosWithWasmClientBootstrapComponents
-{
-    fn governance_proposal_authority(bootstrap: &CosmosWithWasmClientBootstrap) -> &String {
-        &bootstrap.governance_proposal_authority
     }
 }